MdeModulePkg: CodeQL Fixes.
Includes changes across the repo for the following CodeQL rules:
- cpp/comparison-with-wider-type
- cpp/overflow-buffer
- cpp/redundant-null-check-param
- cpp/uselesstest

+++ MdeModulePkg/Universal/FaultTolerantWritePei/FaultTolerantWritePei.c
@@ -102,7 +102,8 @@ FtwGetLastWriteRecord (
   //
   // Try to find the last write record "that has not completed"
   //
-  for (Index = 0; Index < FtwWriteHeader->NumberOfWrites; Index += 1) {
+  for (Index = 0; (UINT64)Index < FtwWriteHeader->NumberOfWrites; Index += 1) {
+    // MU_CHANGE - CodeQL Change - comparison-with-wider-type
     if (FtwRecord->DestinationComplete != FTW_VALID_STATE) {
       //
       // The last write record is found
